changed data structureand  working for all requirements
What changes you have made?
changed my data structure
from object of array objects
to array of objects

Why this change is necessary?
to increase readability and
the old data structure is very hard to read and
understand

What is the impact of the change?
instead of object of array objects now only one array with objects