written a function for arranging input in order
What changes you have made?
written a function for arranging input in order
and changed the code according to it

Why this change is necessary?
before i am using index to get argument value
but it is not a good idea so now arranging them in order and passing
then extracting by order

What is the impact of the change?
now i can access data in an order