Added One more function and if
What changes you have made?
Added one More function and one if to check whether both options are
present

Why this change is necessary?
To validate if both save and query both present

What is the impact of the change?
My app now will give invalid if both options given or non of them are
given